requirements.txt
# Core dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1

# Agent frameworks
langgraph
langchain
langchain-core
langchain-community

# LLM providers
langchain-google-genai
google-generativeai

# Data processing
pandas==2.2.3
numpy==2.2.0
matplotlib==3.10.0
seaborn==0.13.2
plotly==5.24.1

# HTTP client
httpx==0.28.1

# Utilities
aiohttp==3.11.11
beautifulsoup4==4.12.3
lxml==5.3.0
python-multipart==0.0.20
jinja2==3.1.5

# Logging and monitoring
structlog==24.4.0
rich==13.9.4

# Testing dependencies
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-mock==3.14.0
pytest-cov==6.0.0